+++ xenoblade_blender/import_camdo.py
@@ -73,7 +73,7 @@ class ImportCamdo(bpy.types.Operator, ImportHelper):
     ):
         start = time.time()
 
-        database_path = os.path.join(os.path.dirname(__file__), "xcx.json")
+        database_path = os.path.join(os.path.dirname(__file__), "xcx.bin")
         database = xc3_model_py.shader_database.ShaderDatabase.from_file(database_path)
         root = xc3_model_py.load_model_legacy(path, database)